+++ src/capi/frame.cpp
@@ -211,13 +211,35 @@ extern "C" CHFL_PROPERTY* chfl_frame_get_property(const CHFL_FRAME* const frame,
         if (atom_property) {
             property = new Property(*atom_property);
         } else {
-            throw property_error("can not find a property named {} in this frame", name);
+            throw property_error("can not find a property named '{}' in this frame", name);
         }
     )
     return property;
 error:
     delete property;
     return nullptr;
+}
+
+extern "C" chfl_status chfl_frame_add_bond(CHFL_FRAME* const frame, uint64_t i, uint64_t j) {
+    CHECK_POINTER(frame);
+    CHFL_ERROR_CATCH(
+        frame->add_bond(checked_cast(i), checked_cast(j));
+    )
+}
+
+extern "C" chfl_status chfl_frame_remove_bond(CHFL_FRAME* const frame, uint64_t i, uint64_t j) {
+    CHECK_POINTER(frame);
+    CHFL_ERROR_CATCH(
+        frame->remove_bond(checked_cast(i), checked_cast(j));
+    )
+}
+
+extern "C" chfl_status chfl_frame_add_residue(CHFL_FRAME* const frame, const CHFL_RESIDUE* const residue) {
+    CHECK_POINTER(frame);
+    CHECK_POINTER(residue);
+    CHFL_ERROR_CATCH(
+        frame->add_residue(*residue);
+    )
 }
 
 extern "C" chfl_status chfl_frame_free(CHFL_FRAME* const frame) {